requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
httpx==0.27.2
python-dotenv==1.0.1
jinja2==3.1.4
icalendar==6.1.0
recurring-ical-events==3.5.1
